requirements.txt
python-dotenv==1.0.0
python-telegram-bot==13.14
requests==2.28.1